New app with added packages
